refactor: JP_QUERY_MAPの一元化とsession_flavors取得ロジックの共通化
- src/lib/flavor-constants.ts を新規作成し、3ファイルに分散していた
  JP_QUERY_MAP（日本語→英語フレーバー/ブランド名マッピング）を統合
- src/lib/session-service.ts を新規作成し、fetchSessionFlavorsMap 関数を抽出。
  sessions-dashboard, home-session-overview, home-sessions-calendar の
  重複していた session_flavors 取得ロジック（約50行×3箇所）を共通化
- 各コンポーネントは共通モジュールを import するよう変更し、不要な型 import を削除
