+++ src/lib/flavor-constants.ts
@@ -1,0 +1,101 @@
+/**
+ * 日本語 → 英語 フレーバー・ブランド名変換マップ
+ * session-form, home-sessions-calendar, flavors-explorer で共通利用
+ */
+export const JP_QUERY_MAP: Array<{ jp: string; en: string }> = [
+  // フレーバー名
+  { jp: "みんと", en: "mint" },
+  { jp: "ミント", en: "mint" },
+  { jp: "れもん", en: "lemon" },
+  { jp: "レモン", en: "lemon" },
+  { jp: "おれんじ", en: "orange" },
+  { jp: "オレンジ", en: "orange" },
+  { jp: "あっぷる", en: "apple" },
+  { jp: "アップル", en: "apple" },
+  { jp: "りんご", en: "apple" },
+  { jp: "だぶるあっぷる", en: "two apples" },
+  { jp: "ダブルアップル", en: "two apples" },
+  { jp: "ぐれーぷ", en: "grape" },
+  { jp: "グレープ", en: "grape" },
+  { jp: "ぶどう", en: "grape" },
+  { jp: "ぶるーべりー", en: "blueberry" },
+  { jp: "ブルーベリー", en: "blueberry" },
+  { jp: "ちぇりー", en: "cherry" },
+  { jp: "チェリー", en: "cherry" },
+  { jp: "すいか", en: "watermelon" },
+  { jp: "スイカ", en: "watermelon" },
+  { jp: "めろん", en: "melon" },
+  { jp: "メロン", en: "melon" },
+  { jp: "ぴーち", en: "peach" },
+  { jp: "ピーチ", en: "peach" },
+  { jp: "ぱいなっぷる", en: "pineapple" },
+  { jp: "パイナップル", en: "pineapple" },
+  { jp: "ばなな", en: "banana" },
+  { jp: "バナナ", en: "banana" },
+  { jp: "ばにら", en: "vanilla" },
+  { jp: "バニラ", en: "vanilla" },
+  { jp: "ちょこ", en: "chocolate" },
+  { jp: "チョコ", en: "chocolate" },
+  { jp: "しとらす", en: "citrus" },
+  { jp: "シトラス", en: "citrus" },
+  { jp: "らいむ", en: "lime" },
+  { jp: "ライム", en: "lime" },
+  { jp: "ここなっつ", en: "coconut" },
+  { jp: "ココナッツ", en: "coconut" },
+  { jp: "すぱいす", en: "spice" },
+  { jp: "スパイス", en: "spice" },
+  { jp: "でざーと", en: "dessert" },
+  { jp: "デザート", en: "dessert" },
+  // ブランド名
+  { jp: "あるふぁーへる", en: "al fakher" },
+  { jp: "アルファーヘル", en: "al fakher" },
+  { jp: "アルファヘル", en: "al fakher" },
+  { jp: "すたーばず", en: "starbuzz" },
+  { jp: "スターバズ", en: "starbuzz" },
+  { jp: "ふみゃり", en: "fumari" },
+  { jp: "フミャリ", en: "fumari" },
+  { jp: "ふまり", en: "fumari" },
+  { jp: "フマリ", en: "fumari" },
+  { jp: "そしある", en: "social smoke" },
+  { jp: "ソシアル", en: "social smoke" },
+  { jp: "ソーシャル", en: "social smoke" },
+  { jp: "たんじあ", en: "tangiers" },
+  { jp: "タンジア", en: "tangiers" },
+  { jp: "たんじあーず", en: "tangiers" },
+  { jp: "タンジアーズ", en: "tangiers" },
+  { jp: "とりふぇくた", en: "trifecta" },
+  { jp: "トリフェクタ", en: "trifecta" },
+  { jp: "なはら", en: "nakhla" },
+  { jp: "ナハラ", en: "nakhla" },
+  { jp: "ナクラ", en: "nakhla" },
+  { jp: "ロミオワイジュリエット", en: "romeo y julieta" },
+  { jp: "ロミオ", en: "romeo" },
+  { jp: "あずあ", en: "azure" },
+  { jp: "アズア", en: "azure" },
+  { jp: "あずーる", en: "azure" },
+  { jp: "アズール", en: "azure" },
+  { jp: "へいぜ", en: "haze" },
+  { jp: "ヘイズ", en: "haze" },
+  { jp: "あどりあ", en: "adalya" },
+  { jp: "アドリア", en: "adalya" },
+  { jp: "あだりや", en: "adalya" },
+  { jp: "アダリヤ", en: "adalya" },
+  { jp: "さつぃーる", en: "serbetli" },
+  { jp: "サツィール", en: "serbetli" },
+  { jp: "せるべとり", en: "serbetli" },
+  { jp: "セルベトリ", en: "serbetli" },
+  { jp: "せべろ", en: "sebero" },
+  { jp: "セベロ", en: "sebero" },
+  { jp: "だーくさいど", en: "darkside" },
+  { jp: "ダークサイド", en: "darkside" },
+  { jp: "ぶらっくくらうど", en: "black cloud" },
+  { jp: "ブラッククラウド", en: "black cloud" },
+  { jp: "ふーかいん", en: "hookain" },
+  { jp: "フーカイン", en: "hookain" },
+  { jp: "あぐりー", en: "ugly" },
+  { jp: "アグリー", en: "ugly" },
+  { jp: "くっきーず", en: "cookies" },
+  { jp: "クッキーズ", en: "cookies" },
+  { jp: "えたーなるすもーく", en: "eternal smoke" },
+  { jp: "エターナルスモーク", en: "eternal smoke" }
+];
